% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{build_lookup}
\alias{build_lookup}
\title{Initialize a lookup table from a data frame}
\usage{
build_lookup(.data, ...)
}
\arguments{
\item{.data}{a data frame}

\item{...}{column names, tidy}
}
\description{
Build a lookup table rom a data frame according to its column names.
The resulting lookup table is a new data frame.
}
