% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{cover}
\alias{cover}
\title{Initialize a value map from a data frame}
\usage{
cover(.data, ...)
}
\arguments{
\item{.data}{a data frame}

\item{...}{column names, tidy}
}
\description{
Initialize a value map from a data frame
}
