% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{edit_cover}
\alias{edit_cover}
\title{Edit value map}
\usage{
edit_cover(cover, quiet = FALSE)
}
\arguments{
\item{cover}{a value map, or a path pointing to a file containing such}

\item{quiet}{commit the modification without confirmation?}
}
\description{
Edit value map
}
