% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{lookup_location}
\alias{lookup_location}
\title{Location Lookup}
\usage{
lookup_location(
  .str,
  .type = c("country", "city", "state"),
  .match = c("complete", "partial"),
  .min_sim = 0.75,
  .max_return = 5,
  .country_code = NA
)
}
\arguments{
\item{.str}{A string with location Data}

\item{.type}{c("country", "city", "state")}

\item{.match}{c("complete", "partial")}

\item{.min_sim}{Minimum OSA similarity for fuzzy matches (default = .75)}

\item{.max_return}{Maximum Number of returned matches (default = 5)}

\item{.country_code}{Country to search the in}
}
\value{
A Dataframe
}
\description{
Location Lookup
}
\examples{
.str        <- c("MARTINEZ, CA,")
.type       <- c("country", "city", "state")
.max_dist   <- 2
.match      <- "partial"
.min_sim    <- .75
.max_return <- 5
}
