% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{modify_cover}
\alias{modify_cover}
\title{Batch modify a value map}
\usage{
modify_cover(cover, .fun, ...)
}
\arguments{
\item{cover}{a value map}

\item{.fun}{a function to be applied to old_value}

\item{...}{a vector containing col_name}
}
\description{
Batch modify a value map
}
