% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{package}
\name{rlookup-package}
\alias{rlookup}
\alias{rlookup-package}
\title{rlookup}
\description{
Recoding variables in a data frame through a value map, through making a surjection of the old_value to the new_value
}
