% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{standardize_country}
\alias{standardize_country}
\title{Wrapper around lookup_location}
\usage{
standardize_country(.str, .min_sim = 0.25, .sim = FALSE)
}
\arguments{
\item{.str}{A character string with country names}

\item{.min_sim}{Minimum similarity}

\item{.sim}{Include similarity in character names}
}
\value{
A charcter with ISO 3 codes
}
\description{
Wrapper around lookup_location
}
\examples{
standardize_country(c("Palestine", "Palestine", "Palestine"), .1, .sim = TRUE)

}
