% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{standardize_names}
\alias{standardize_names}
\title{Standardize Names}
\usage{
standardize_names(.str)
}
\arguments{
\item{.str}{A character String}
}
\value{
A character String
}
\description{
Standardize Names
}
