% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{use_cover}
\alias{use_cover}
\title{Use value map to replace values in a data frame}
\usage{
use_cover(.data, cover, mark = c("old", "new", "both"),
  drop_old = TRUE)
}
\arguments{
\item{.data}{a data frame}

\item{cover}{a value map, or a path pointing to a file containing such}

\item{mark}{out of old and new, mark which column name? choose from: c("old", "new", "both")}

\item{drop_old}{keep only the new column after recoding}
}
\description{
Use value map to replace values in a data frame
}
