% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{use_lookup}
\alias{use_lookup}
\title{Use lookup table to replace values in a data frame}
\usage{
use_lookup(.data, lookup, ..., drop_old = TRUE, mark = c("old", "new",
  "both"), old_suffix = "__old__", new_suffix = "__new__",
  use_na = FALSE)
}
\arguments{
\item{.data}{a data frame}

\item{lookup}{a lookup table, or a path pointing to a file containing such}

\item{...}{column names}

\item{drop_old}{keep only the new column after recoding}

\item{mark}{mark which column name in the resulting data frame? choose from: "old", "new" or "both"}

\item{old_suffix}{a string for suffixing column containing old values}

\item{new_suffix}{a string for suffixing column containing new values}

\item{use_na}{use NAs in `new_value` for replacement; if FALSE, use `old_value` as `new_value`}
}
\description{
Use lookup table to replace values in a data frame
}
