% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{write_cover}
\alias{write_cover}
\title{Write value map to a file}
\usage{
write_cover(cover, path, overwrite = FALSE, edit = FALSE)
}
\arguments{
\item{cover}{a value map}

\item{path}{a character string}

\item{overwrite}{overwriting existing file from \code{path}}

\item{edit}{invoke \code{edit_cover} before writing to file}
}
\description{
Write value map to a file
}
