% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{write_lookup}
\alias{write_lookup}
\title{Write lookup table to a file}
\usage{
write_lookup(lookup, path, overwrite = FALSE)
}
\arguments{
\item{lookup}{a lookup table}

\item{path}{a character string}

\item{overwrite}{overwriting existing file from \code{path}}
}
\description{
Write lookup table to a file
}
