% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{edit_lookup}
\alias{edit_lookup}
\title{Edit value map}
\usage{
edit_lookup(lookup, quiet = FALSE)
}
\arguments{
\item{lookup}{a value map, or a path pointing to a file containing such}

\item{quiet}{commit the modification without confirmation?}
}
\description{
Changes made on an object yields a new copy; Changes made on a file are written back to the file.
}
