% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{modify_lookup}
\alias{modify_lookup}
\title{Batch modify a value map}
\usage{
modify_lookup(lookup, .fun, ...)
}
\arguments{
\item{lookup}{a value map}

\item{.fun}{a function to be applied to old_value}

\item{...}{a vector containing col_name}
}
\description{
Apply a function on `old_value`s and assign the result to `new_value`s.
Changes made on an object yields a new copy; Changes made on a file are written back to the file.
}
