% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{write_lookup}
\alias{write_lookup}
\title{Write value map to a file}
\usage{
write_lookup(lookup, path, overwrite = FALSE, edit = FALSE)
}
\arguments{
\item{lookup}{a value map}

\item{path}{a character string}

\item{overwrite}{overwriting existing file from \code{path}}

\item{edit}{invoke \code{edit_lookup} before writing to file}
}
\description{
Write value map to a file
}
